+++ CHANGELOG.md
@@ -76,6 +76,16 @@ configuration property (refer [examples/rdkafka_complex_consumer_example.c]
    is now built-in. There are bundled builds with and without linking to
    cyrus-sasl for maximum compatibility.
  * Added `rd_kafka_topic_partition_get_leader_epoch()` (and `set..()`).
+ * Added partition leader epoch APIs:
+   - `rd_kafka_topic_partition_get_leader_epoch()` (and `set..()`)
+   - `rd_kafka_message_leader_epoch()`
+   - `rd_kafka_*assign()` and `rd_kafka_seek_partitions()` now supports
+     partitions with a leader epoch set.
+   - `rd_kafka_offsets_for_times()` will return per-partition leader-epochs.
+   - `leader_epoch`, `stored_leader_epoch`, and `committed_leader_epoch`
+     added to per-partition statistics.
+
+
 
 ## Fixes